smartcontracts page created and layout of it is adjusted.
