+++ profile/website/src/components/index.js
@@ -4,3 +4,4 @@ export { default as Footer } from './Footer';
 export { default as Services } from './Services';
 export { default as Transactions } from './Transactions';
 export { default as Home } from './Home';
+export { default as Smartcontracts} from './Smartcontracts';
